requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Parsing
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.10

# Image Processing
pillow==10.2.0

# AI Integration
openai==1.10.0

# Telegram
aiogram==3.4.1

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Scheduler
apscheduler==3.10.4

# Logging
loguru==0.7.2
